web: Document the labels.  Fix the alerts
This commit documents and adds unit tests for our `Labels` component, which are usually called
"chips" in other design systems.

I've also reverted to allowing the components that take 'level' information to take it as a single
argument that is _either_ an attribute or a property.  If it's a property, it reverts to the older
behavior.