web: fix early modal stack depletion
While working on something else, I discovered this mutation inside
the modal stack handler; I've changed it to make a copy of the
original stack, modify that copy, and then write the changed stack
back to the original.  While unlikely, it is possible that the stack
could get out-of-sync and be depleted before all stacked modals
have been closed.  This fixes that issue.