web: add 'remember me' feature to login
Provides a checkbox labeled "Remember me on this device" below the Username field of the
IdentificationStage log in. When checked, the username entered will be saved in the user's
LocalStorage. When unchecked, any LocalStorage fields related to the `remember-me` feature will be
found and deleted.

When reaching the IdentificationStage with the `remember-me` fields in LocalStorage activate,
populated, and coherent, the username will be automatically filled in. If the password field is
present on the IdentificationStage dialog, the user will remain here; if it is not, the server will
automatically be triggered to proceed to the next stage. If, in the event that the server rejects
the username (which can happen if `pretend users exists` is set to False) and the user is returned
to the IdentificationStage, this feature must not cause an infinite loop.

If the user is on the PasswordStage and presses the "Not you?" anchor, any LocalStorage fields
related to the `remember-me` feature will be found and deleted before the user is sent back to the
previous stage.

The *Update Identification Stage* admin dialog has a new toggle, **Enable "Remember me on this
device"**, to allow administrators to enable or disable this feature on a per-flow basis.

To be as unobtrusive as possible, the "remember me" logic has been isolated into a specialized
controller for IdentificationStage.  This controller uses only a few touchpoints within the
IdentificationStage, rather than being a scattered chunk of logic throughout.

Logic in the Controller:

- On `connectedCallback,` it checks if this feature is enabled and does a sanity check to prevent an
  infinite loop on auth failure. If the sanity check fails, it resets this feature for the current
  user.
- On `willUpdate(),` it attempts to identify the username and provide a value for the client.
- on `updated(),` it checks if the username is filled in, and has been filled in by us, and there are
  no password fields or other auth methods visible. If all this is true, it submits the username to
  the server immediately, moving to the Password stage.
- on `render()`, if enabled, it provides the checkbox dialog.
- on `click`:
  - if enabled: records current username and sets up listeners to record changes to username while
    the user is here.
  - if disabled: clears and deletes users and removes any existing listeners.

Touchpoints in IdentificationStage (each of these is one line of code):

- Include the controller
- Include styles provided by the controller for its visual component
- Initialize the controller
- Consult the controller for any available username
- Render the controller's visual component.

Requested by several users. Seems to be a popular feature. Reaches for parity with competitors.