website: add links to schema references
Added links to the schema references on the blueprints page, and a new collapsible
section that shows all the past schemas as referenced in the original edit.

URLs to live objects (URLs that aren't just examples) should be
live links themselves.  Referring to past versions of the documents
and not providing a list of those versions or links to them seemed
incomplete.